refactor(tune): H20 terminal form: the policy file owns the seed heuristic
Follows upstream's policies/seed-selection revision: the former legacy
get_config body moves into sm90_h20_families as build_h20_seed_config (the
single implementation for all dtypes); sm90_h20 becomes a thin dispatcher.
Zero-drift verified on a 1644-point pre/post snapshot spanning both parity
grids and the out-of-scope dtype universes.
